backend/requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
awsebcli==3.14.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
blessed==1.14.2
botocore==1.10.4
cached-property==1.5.1
cement==2.8.2
click==6.7
colorama==0.3.9
decorator==4.3.0
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
Flask==1.0.2
Flask-RESTful==0.3.6
funcsigs==1.0.2
functools32==3.2.3.post2
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.3.0
pathlib2==2.3.2
pathspec==0.5.5
pluggy==0.7.1
psutil==5.4.6
py==1.6.0
pyte==0.8.0
pytest==3.8.0
python-dateutil==2.6.1
pytz==2018.5
PyYAML==3.12
requests==2.9.1
scandir==1.7
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
thefuck==3.27
virtualenv==16.0.0
wcwidth==0.1.7
websocket-client==0.47.0
Werkzeug==0.14.1
